fix: resolve all payout and bankroll bugs, refactor game engine for production accuracy
